Update workly.js with new custom function only applicable to new task creations.
By creating a new function called 'addNewTask' that is ONLY applicable to a new task and not rendering all existing tasks (as done on app startup), the app successfully shows the name of the task. I spent an entire night on this bug when I created workly, but now workly works (for the most part).